Move palette selector to separate component
